requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.1
djangorestframework==3.14.0
drf-yasg==1.21.7
greenlet==3.0.1
idna==3.6
inflection==0.5.1
opengrapher==0.4
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyMySQL==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
SQLAlchemy==2.0.23
sqlparse==0.4.4
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
